+++ src/seeders/20220309202416-User.js
@@ -1,37 +1,26 @@
+'use strict';
 const bcrypt = require('bcrypt');
 const salt = bcrypt.genSaltSync(10, 'b');
-module.exports = {
-	up: (queryInterface, Sequelize) =>
-		queryInterface.bulkInsert(
-			'Users',
-			[
 
-				{
-					names:'YANGENEYE Patrick',
-					email:'yangeney@gmail.com',
-					password:await bcrypt.hash('password', salt),
-					roleId:1,
-					createdAt: new Date(),
-			     	updatedAt: new Date()
-				},
-				{
-					names:'seeduser',
-  					email:'seeduser@gmail.com',
-    				password:await bcrypt.hash('seeduser', salt),
-					roleId:5,
-					createdAt: new Date(),
-					updatedAt: new Date()
-				},
-				{
-					names: 'Gihozo Innocente',
-					email: 'gihozo97@gmail.com',
-					password:
-						'$2b$10$UXnwEWSIsSXsK/ocfmr0R.7OAt3WD1WpD0kx5VtI/wNbvQO1iGxWa',
-					managerId: 1,
-					createdAt: new Date(),
-					updatedAt: new Date(),
-				},
-				{
+module.exports = {
+  async up (queryInterface, Sequelize) {
+      await queryInterface.bulkInsert('Users', [{
+      names:'YANGENEYE Patrick',
+      email:'yangeney@gmail.com',
+      password:await bcrypt.hash('password', salt),
+        roleId:1,
+        createdAt: new Date(),
+        updatedAt: new Date()
+      },
+      {
+        names:'seeduser',
+        email:'seeduser@gmail.com',
+        password:await bcrypt.hash('seeduser', salt),
+          roleId:5,
+          createdAt: new Date(),
+          updatedAt: new Date()
+        },
+        {
 					names: 'Honore Iradukunda',
 					email: 'ihonore03@gmail.com',
 					password:
@@ -60,12 +49,12 @@ module.exports = {
 					managerId: 1,
 					createdAt: new Date(),
 					updatedAt: new Date(),
-				},
-			],
+				},], {});
+    
+  },
 
-			{}
-		),
-
-	down: (queryInterface, Sequelize) =>
-		queryInterface.bulkDelete('Users', null, {}),
+  async down (queryInterface, Sequelize) {
+      await queryInterface.bulkDelete('Users', null, {});
+     
+  }
 };
